refactor(ui): 重构历史记录页面结构，抽离为独立目录模块
