+++ src/danmaku_sender/ui/history/__init__.py
@@ -1,0 +1,4 @@
+from .page import HistoryPage
+
+
+__all__ = ["HistoryPage"]
